+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Bumped dependency PowershellGet to PSResourceGet
+
 ## [1.4.1] - 2023-08-23
 
 ### Fixed