+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2024-03-27
+
 ### Added
 
 - Compare-ObjectProperties; Added new command to compare all properties of two objects